Support []byte in String and FixedString
A common use case in ClickHouse is to store json in String data.
Golang functions that marshal json return a []byte rather than a
string.

The string type has a few behaviors attached that are handy for dealing
with UTF-8 text, but it'll hold whatever bytes you put in it.

Text-handling stuff in the standard library is often written to work with
[]bytes too, e.g. package bytes mirrors package strings and regexp deals in
either.